requirements.txt
djangorestframework
django
drf-yasg
psycopg2-binary
psycopg2
djongo